HDDS-5371. [FSO] Remove all the dependencies on OzoneManager#getEnableFileSystemPaths.
