+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/request/OMClientRequest.java
@@ -484,10 +484,11 @@ public abstract class OMClientRequest implements RequestAuditor {
     }
   }
 
-  public static String validateAndNormalizeKey(boolean enableFileSystemPaths,
-      String keyPath, BucketLayout bucketLayout) throws OMException {
+  public static String validateAndNormalizeKey(String keyPath,
+                                               BucketLayout bucketLayout)
+      throws OMException {
     LOG.debug("Bucket Layout: {}", bucketLayout);
-    if (bucketLayout.shouldNormalizePaths(enableFileSystemPaths)) {
+    if (bucketLayout.isFileSystemOptimized()) {
       keyPath = validateAndNormalizeKey(true, keyPath);
       if (keyPath.endsWith("/")) {
         throw new OMException(
