+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/response/TestCleanupTableInfo.java
@@ -185,7 +185,6 @@ public class TestCleanupTableInfo {
   @Test
   public void testKeyCreateRequestSetsAllTouchedTableCachesForEviction() {
     OMKeyCreateRequest request = anOMKeyCreateRequest();
-    when(om.getEnableFileSystemPaths()).thenReturn(true);
 
     Map<String, Integer> cacheItemCount = recordCacheItemCounts();
